mb/gigabyte/ga-b75m-d3v: add libgfxinit support
On this device libgfxinit successfully initializes textmode and 
framebuffer mode on the device connected to the HDMI1 port.

VGA port is not tested, due to lack of VGA with valid EDID.

Every ports works on GNU/Linux with kernel version 4.9+80+deb9u1.

Due to limitations within the libgfxinit's current build system,
it needs "select NORTHBRIDGE_INTEL_IVYBRIDGE" to work, even it is
PCH instead of CPU for which libgfxinit should be configured here.

Change-Id: Ib1513fc57b79eebc5d047c2e24585731c44bc215
Signed-off-by: Bill XIE <persmule@gmail.com>
Reviewed-on: https://review.coreboot.org/21071
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>